% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcast.R
\name{funcast}
\alias{funcast}
\title{Create a FunCast model}
\usage{
funcast(
  K = 10L,
  s = 0.5,
  basis_type = "bspline",
  auto_h = TRUE,
  h_list = NULL,
  degree = 3L,
  rcond = NULL
)
}
\arguments{
\item{K}{Number of basis functions for the future of Y.}

\item{s}{Smoothing coefficient.}

\item{basis_type}{"bspline" or "fourier".}

\item{auto_h}{If TRUE, h is selected with RRSS.}

\item{h_list}{Values of h if \code{auto_h = FALSE}.}

\item{degree}{B-spline degree.}

\item{rcond}{Pseudo-inverse threshold.}
}
\value{
An object of class \code{funcast}.
}
\description{
Create a FunCast model
}
