% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.funcast}
\alias{predict.funcast}
\title{Predict with a FunCast model}
\usage{
\method{predict}{funcast}(object, Y_past_new, covariates_past_new = NULL, ...)
}
\arguments{
\item{object}{A fitted \code{funcast} object.}

\item{Y_past_new}{Matrix, n_new rows and length(t_past) columns.}

\item{covariates_past_new}{List of matrices, same shape as \code{Y_past_new}.}

\item{...}{Unused.}
}
\value{
Matrix, n_new rows and length(t_future) columns.
}
\description{
Predict with a FunCast model
}
